No real changes
Changed pathfinding to take floats, doesn't change anything really.
Mostly experimental.

+++ Clyde.cpp
@@ -12,8 +12,8 @@ Clyde::Clyde(FXMVECTOR pos, float radius) : Ghost(pos, radius)
 	this->mChaseTimer = 0;
 
 	//Draw the path to his scatter area prior to the start of the game to prevent bottlenecks
-	mStart = new PathNode((int)this->mPos.x, (int)this->mPos.z);
-	mGoal = new PathNode((int)this->mScatterTile.x, (int)this->mScatterTile.z);
+	mStart = new PathNode(this->mPos.x, this->mPos.z);
+	mGoal = new PathNode(this->mScatterTile.x, this->mScatterTile.z);
 	mWaypoints = path.FindPath(mStart, mGoal);
 	scatterPathDrawn = true;
 }
@@ -24,40 +24,40 @@ Clyde::~Clyde()
 
 void Clyde::LoadScatterWaypoints()
 {
-	mScatterWaypoints.push_back(new PathNode((int)-13.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-12.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-11.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-10.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-9.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-8.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-7.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-6.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-5.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-4.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-3.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-2.0f, (int)-13.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-2.0f, (int)-12.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-2.0f, (int)-11.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-2.0f, (int)-10.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-3.0f, (int)-10.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-4.0f, (int)-10.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-5.0f, (int)-10.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-5.0f, (int)-9.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-5.0f, (int)-8.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-5.0f, (int)-7.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-6.0f, (int)-7.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-7.0f, (int)-7.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-8.0f, (int)-7.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-8.0f, (int)-8.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-8.0f, (int)-9.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-8.0f, (int)-10.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-9.0f, (int)-10.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-10.0f, (int)-10.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-11.0f, (int)-10.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-12.0f, (int)-10.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-13.0f, (int)-10.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-13.0f, (int)-11.5f));
-	mScatterWaypoints.push_back(new PathNode((int)-13.0f, (int)-12.5f));
+	mScatterWaypoints.push_back(new PathNode(-13.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-12.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-11.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-10.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-9.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-8.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-7.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-6.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-5.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-4.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-3.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-2.0f, -13.5f));
+	mScatterWaypoints.push_back(new PathNode(-2.0f, -12.5f));
+	mScatterWaypoints.push_back(new PathNode(-2.0f, -11.5f));
+	mScatterWaypoints.push_back(new PathNode(-2.0f, -10.5f));
+	mScatterWaypoints.push_back(new PathNode(-3.0f, -10.5f));
+	mScatterWaypoints.push_back(new PathNode(-4.0f, -10.5f));
+	mScatterWaypoints.push_back(new PathNode(-5.0f, -10.5f));
+	mScatterWaypoints.push_back(new PathNode(-5.0f, -9.5f));
+	mScatterWaypoints.push_back(new PathNode(-5.0f, -8.5f));
+	mScatterWaypoints.push_back(new PathNode(-5.0f, -7.5f));
+	mScatterWaypoints.push_back(new PathNode(-6.0f, -7.5f));
+	mScatterWaypoints.push_back(new PathNode(-7.0f, -7.5f));
+	mScatterWaypoints.push_back(new PathNode(-8.0f, -7.5f));
+	mScatterWaypoints.push_back(new PathNode(-8.0f, -8.5f));
+	mScatterWaypoints.push_back(new PathNode(-8.0f, -9.5f));
+	mScatterWaypoints.push_back(new PathNode(-8.0f, -10.5f));
+	mScatterWaypoints.push_back(new PathNode(-9.0f, -10.5f));
+	mScatterWaypoints.push_back(new PathNode(-10.0f, -10.5f));
+	mScatterWaypoints.push_back(new PathNode(-11.0f, -10.5f));
+	mScatterWaypoints.push_back(new PathNode(-12.0f, -10.5f));
+	mScatterWaypoints.push_back(new PathNode(-13.0f, -10.5f));
+	mScatterWaypoints.push_back(new PathNode(-13.0f, -11.5f));
+	mScatterWaypoints.push_back(new PathNode(-13.0f, -12.5f));
 }
 
 void Clyde::Update(float dt, bool powerUpActivated, int levelNumber, int pelletCounter)
@@ -82,11 +82,11 @@ void Clyde::Update(float dt, bool powerUpActivated, int levelNumber, int pelletC
 	if (!isIdle)
 	{
 		//Get clyde's position in row/col space
-		/*int clydePos = ((MazeLoader::GetMazeHeight()) - (int)round(this->getPos().z + 15.5f)) +
-			((MazeLoader::GetMazeWidth()) - (int)round(this->getPos().x + 15.5f));
+		/*int clydePos = ((MazeLoader::GetMazeHeight()) - round(this->getPos().z + 15.5f)) +
+			((MazeLoader::GetMazeWidth()) - round(this->getPos().x + 15.5f));
 		//Get puckman's position in row/col space
-		int puckmanPos = ((MazeLoader::GetMazeHeight()) - (int)round(MazeLoader::GetPacManData().at(0).pos.x + 15.5f)) +
-			((MazeLoader::GetMazeWidth()) - (int)round(MazeLoader::GetPacManData().at(0).pos.z + 15.5f));
+		int puckmanPos = ((MazeLoader::GetMazeHeight()) - round(MazeLoader::GetPacManData().at(0).pos.x + 15.5f)) +
+			((MazeLoader::GetMazeWidth()) - round(MazeLoader::GetPacManData().at(0).pos.z + 15.5f));
 		//Determine the euclidean distance between clyde and puckman
 		int euclidDistance = abs(clydePos - puckmanPos);
 		//If the distance between Clyde and PuckMan is 8 tiles or more in Euclidean space, target PuckMan and chase like Blinky does
@@ -112,8 +112,8 @@ void Clyde::Update(float dt, bool powerUpActivated, int levelNumber, int pelletC
 		case SCATTER:
 			if (!scatterPathDrawn)
 			{
-				mStart = new PathNode((int)this->mPos.x, (int)this->mPos.z);
-				mGoal = new PathNode((int)this->mScatterTile.x, (int)this->mScatterTile.z);
+				mStart = new PathNode(this->mPos.x, this->mPos.z);
+				mGoal = new PathNode(this->mScatterTile.x, this->mScatterTile.z);
 				mWaypoints = path.FindPath(mStart, mGoal);
 				this->SetWayPoints(mWaypoints);
 				scatterPathDrawn = true;
@@ -176,20 +176,20 @@ void Clyde::Update(float dt, bool powerUpActivated, int levelNumber, int pelletC
 		case CHASE:
 			if (!firstChasePathDrawn)
 			{
-				mStart = new PathNode((int)this->mPos.x, (int)this->mPos.z);
-				mGoal = new PathNode((int)round(MazeLoader::GetPacManData().at(0).pos.x), (int)round(MazeLoader::GetPacManData().at(0).pos.z));
+				mStart = new PathNode(this->mPos.x, this->mPos.z);
+				mGoal = new PathNode(round(MazeLoader::GetPacManData().at(0).pos.x), round(MazeLoader::GetPacManData().at(0).pos.z));
 				mWaypoints = path.FindPath(mStart, mGoal);
 				waypointIterator = 0;
 				firstChasePathDrawn = true;
 			}
 			else
 			{
-				int row = MazeLoader::GetMazeHeight() - (int)round(this->mPos.x + 15.5f);
-				int col = (int)round(this->mPos.z + 14.5f) - 1;
+				int row = MazeLoader::GetMazeHeight() - round(this->mPos.x + 15.5f);
+				int col = round(this->mPos.z + 14.5f) - 1;
 				if (MazeLoader::IsDivergent(row, col))
 				{
-					mStart = new PathNode((int)this->mPos.x, (int)this->mPos.z);
-					mGoal = new PathNode((int)round(MazeLoader::GetPacManData().at(0).pos.x), (int)round(MazeLoader::GetPacManData().at(0).pos.z));
+					mStart = new PathNode(this->mPos.x, this->mPos.z);
+					mGoal = new PathNode(round(MazeLoader::GetPacManData().at(0).pos.x), round(MazeLoader::GetPacManData().at(0).pos.z));
 					mWaypoints = path.FindPath(mStart, mGoal);
 					waypointIterator = 0;
 				}
